Refactoring file path
