+++ lib/diagram/common/tooltip-creator.js
@@ -16,40 +16,40 @@ import {
   OpsLinkAttribute,
   OpsNodeAttribute,
   OpsTermPointAttribute
-} from '../../../server/graph/rfc-model/model-ops-attr'
-import RfcL2LinkAttribute from '../../../server/graph/rfc-model/link-attr-rfc-l2'
-import RfcL3LinkAttribute from '../../../server/graph/rfc-model/link-attr-rfc-l3'
-import RfcL2NetworkAttribute from '../../../server/graph/rfc-model/network-attr-rfc-l2'
-import RfcL3NetworkAttribute from '../../../server/graph/rfc-model/network-attr-rfc-l3'
-import RfcL2NodeAttribute from '../../../server/graph/rfc-model/node-attr-rfc-l2'
-import RfcL3NodeAttribute from '../../../server/graph/rfc-model/node-attr-rfc-l3'
-import RfcL2TermPointAttribute from '../../../server/graph/rfc-model/term-point-attr-rfc-l2'
-import RfcL3TermPointAttribute from '../../../server/graph/rfc-model/term-point-attr-rfc-l3'
-import AggregatedNodeAttribute from '../../../server/graph/rfc-model/node-attr-aggr'
+} from '../../../server/graph/rfc-model/model/ops-attr'
+import RfcL2LinkAttribute from '../../../server/graph/rfc-model/link-attr/rfc-l2'
+import RfcL3LinkAttribute from '../../../server/graph/rfc-model/link-attr/rfc-l3'
+import RfcL2NetworkAttribute from '../../../server/graph/rfc-model/network-attr/rfc-l2'
+import RfcL3NetworkAttribute from '../../../server/graph/rfc-model/network-attr/rfc-l3'
+import RfcL2NodeAttribute from '../../../server/graph/rfc-model/node-attr/rfc-l2'
+import RfcL3NodeAttribute from '../../../server/graph/rfc-model/node-attr/rfc-l3'
+import RfcL2TermPointAttribute from '../../../server/graph/rfc-model/tp-attr/rfc-l2'
+import RfcL3TermPointAttribute from '../../../server/graph/rfc-model/tp-attr/rfc-l3'
+import AggregatedNodeAttribute from '../../../server/graph/rfc-model/node-attr/aggr'
 import {
   MddoL1NetworkAttribute,
   MddoL2NetworkAttribute,
   MddoL3NetworkAttribute,
   MddoOspfAreaNetworkAttribute
-} from '../../../server/graph/rfc-model/network-attr-mddo'
+} from '../../../server/graph/rfc-model/network-attr/mddo'
 import {
   MddoL1NodeAttribute,
   MddoL2NodeAttribute,
   MddoL3NodeAttribute,
   MddoOspfAreaNodeAttribute
-} from '../../../server/graph/rfc-model/node-attr-mddo'
+} from '../../../server/graph/rfc-model/node-attr/mddo'
 import {
   MddoL1LinkAttribute,
   MddoL2LinkAttribute,
   MddoL3LinkAttribute,
   MddoOspfAreaLinkAttribute
-} from '../../../server/graph/rfc-model/link-attr-mddo'
+} from '../../../server/graph/rfc-model/link-attr/mddo'
 import {
   MddoL1TermPointAttribute,
   MddoL2TermPointAttribute,
   MddoL3TermPointAttribute,
   MddoOspfAreaTermPointAttribute
-} from '../../../server/graph/rfc-model/term-point-attr-mddo'
+} from '../../../server/graph/rfc-model/tp-attr/mddo'
 
 /**
  * Attribute class string to class dictionary.
