+++ server/graph/nested/aggregated-node.js
@@ -2,7 +2,7 @@
  * @file Definition of aggregated Node.
  */
 
-import AggregatedNodeAttribute from '../rfc-model/node-attr-aggr.js'
+import AggregatedNodeAttribute from '../rfc-model/node-attr/aggr.js'
 import DeepNestedNode from './deep-node'
 
 /**
