+++ server/graph/rfc-model/link-attr/rfc-l2.js
@@ -2,7 +2,7 @@
  * @file Attribute class for layer2 link of topology model.
  */
 
-import RfcModelBase from './base'
+import RfcModelBase from '../base'
 
 /**
  * Attribute class for layer2 link.
