+++ server/graph/rfc-model/link-attr/rfc-l3.js
@@ -2,7 +2,7 @@
  * @file Attribute class for layer3 link of topology model.
  */
 
-import RfcModelBase from './base'
+import RfcModelBase from '../base'
 
 /**
  * Attribute class for layer3 link.
