+++ server/graph/rfc-model/network-attr/rfc-l3.js
@@ -2,7 +2,7 @@
  * @file Attribute class for layer3 network of topology model.
  */
 
-import RfcModelBase from './base'
+import RfcModelBase from '../base'
 
 /**
  * Attribute class for layer3 network.
