+++ server/graph/rfc-model/node-attr/rfc-l2.js
@@ -1,7 +1,7 @@
 /**
  * @file Attribute class for layer2 node of topology model.
  */
-import RfcModelBase from './base'
+import RfcModelBase from '../base'
 
 /**
  * Attribute class for layer2 node.
