+++ server/graph/rfc-model/tp-attr/rfc-l2.js
@@ -2,7 +2,7 @@
  * @file Attribute class for layer2 term-point of topology model.
  */
 
-import RfcModelBase from './base'
+import RfcModelBase from '../base'
 
 /**
  * VLAN ID/Name class.
